Refactor auth into shared context provider
